# WIP: Performance of committee kernels 