This fixes a type-instability deep inside the commitee kernels 
```
BEFORE FIX 

(base) ortner@CosmicCrisp profile % j18 --project=.. -t 8 profile_committees.jl
nthreads() = 8
len = length(basis) = 1119
length(at) = 256
[ Info: Energy
Potential:   0.004407 seconds (240.58 k allocations: 7.521 MiB)
Committee:   0.095018 seconds (5.43 M allocations: 719.990 MiB, 45.24% gc time)
[ Info: Forces
Potential:   0.025057 seconds (677.80 k allocations: 21.844 MiB)
Committee:  29.404100 seconds (2.46 G allocations: 168.960 GiB, 48.25% gc time)
[ Info: Virial
Potential:   0.025030 seconds (677.78 k allocations: 21.749 MiB)
Committee:  29.351164 seconds (2.46 G allocations: 168.959 GiB, 48.24% gc time)



AFTER FIX

(base) ortner@CosmicCrisp profile % j18 --project=.. -t 8 profile_committees.jl
nthreads() = 8
len = length(basis) = 1119
length(at) = 256
[ Info: Energy
Potential:   0.004448 seconds (240.44 k allocations: 7.518 MiB)
Committee:   0.005977 seconds (240.46 k allocations: 7.517 MiB)
[ Info: Forces
Potential:   0.026286 seconds (677.43 k allocations: 21.834 MiB)
Committee:   0.153319 seconds (679.02 k allocations: 207.272 MiB, 43.56% gc time)
[ Info: Virial
Potential:   0.027093 seconds (677.41 k allocations: 21.739 MiB)
Committee:   0.092613 seconds (679.02 k allocations: 205.809 MiB, 10.13% gc ti
```

New results look better but TBH far from great. I know the source of the slow-down and thought I had a fix for it, but it does not yet work as expected.

Should this still be merged and tagged? 

CC @casv2 @wcwitt