@wcwitt 2022-12-13
Here is what I get for the example I reported initially (#52):

```
BEFORE

julia> include("test.jl")
Time for forces:  0.008954724
Time for committee forces:  6.388195141

AFTER

julia> include("test.jl")
Time for forces:  0.008372433
Time for committee forces:  0.066866573
```

So I would vote for merging and tagging. Thank you!