@cortner 2022-12-13
This adresses #52 and is now registered as 0.10.3.